MAGE-938 Add more robust handling of unknown unused replica indices with CLI confirmation as these deletions are potentially risky

+++ Console/Command/ReplicaDeleteCommand.php
@@ -3,9 +3,8 @@
 namespace Algolia\AlgoliaSearch\Console\Command;
 
 use Algolia\AlgoliaSearch\Api\Product\ReplicaManagerInterface;
+use Algolia\AlgoliaSearch\Exceptions\BadRequestException;
 use Algolia\AlgoliaSearch\Service\StoreNameFetcher;
-use Magento\Framework\App\Area;
-use Magento\Framework\App\State;
 use Magento\Framework\Console\Cli;
 use Magento\Store\Model\StoreManagerInterface;
 use Symfony\Component\Console\Command\Command;
@@ -13,6 +12,7 @@ use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Question\ConfirmationQuestion;
 
 class ReplicaDeleteCommand extends Command
 {
@@ -22,9 +22,9 @@ class ReplicaDeleteCommand extends Command
     protected const UNUSED_OPTION_SHORTCUT = 'u';
 
     protected ?OutputInterface $output = null;
+    protected ?InputInterface $input = null;
 
     public function __construct(
-        protected State                   $state,
         protected ReplicaManagerInterface $replicaManager,
         protected StoreManagerInterface   $storeManager,
         protected StoreNameFetcher        $storeNameFetcher,
@@ -71,13 +71,69 @@ class ReplicaDeleteCommand extends Command
         }
 
         $this->output = $output;
-//        $this->state->setAreaCode(Area::AREA_ADMINHTML);
+        $this->input = $input;
 
-        $this->deleteReplicas($storeIds, $unused);
+        if ($unused) {
+            $unusedReplicas = $this->getUnusedReplicas($storeIds);
+            if (!$unusedReplicas) {
+                $output->writeln('<comment>No unused replicas found.</comment>');
+                return Cli::RETURN_SUCCESS;
+            }
+            if (!$this->confirmDeleteUnused($unusedReplicas)) {
+                return Cli::RETURN_SUCCESS;
+            }
+        }
+
+        try {
+            $this->deleteReplicas($storeIds, $unused);
+        } catch (BadRequestException $e) {
+            $this->output->writeln("<error>Error encountered while attempting to delete replica: {$e->getMessage()}</error>");
+            $this->output->writeln('<comment>It is likely that the Magento integration does not manage the index. You should review your application configuration in Algolia.</comment>');
+            return CLI::RETURN_FAILURE;
+        }
 
         return Cli::RETURN_SUCCESS;
     }
 
+    protected function getUnusedReplicas(array $storeIds): array
+    {
+        return array_reduce(
+            $storeIds,
+            function($allUnused, $storeId) {
+                $unused = [];
+                try {
+                    $unused = $this->replicaManager->getUnusedReplicaIndices($storeId);
+                } catch (\Exception $e) {
+                    $this->output->writeln("<error>Unable to retrieve unused replicas for $storeId: {$e->getMessage()}</error>");
+                }
+                return array_unique(array_merge($allUnused, $unused));
+            },
+            []
+        );
+    }
+
+    /**
+     * Deleting unused replica indices is potentially risky, especially if they have enabled query suggestions on their index
+     * Verify with the end user first!
+     *
+     * @param array $unusedReplicas
+     * @return bool
+     */
+    protected function confirmDeleteUnused(array $unusedReplicas): bool
+    {
+        $this->output->writeln('<info>The following replicas appear to be unused and will be deleted:</info>');
+        foreach ($unusedReplicas as $unusedReplica) {
+            $this->output->writeln('<info> - ' . $unusedReplica . '</info>');
+        }
+        $helper = $this->getHelper('question');
+        $question = new ConfirmationQuestion('<question>Do you want to proceed? (y/n)</question> ', false);
+
+        if (!$helper->ask($this->input, $this->output, $question)) {
+            $this->output->writeln('<comment>Operation cancelled.</comment>');
+            return false;
+        }
+        return true;
+    }
 
     protected function deleteReplicas(array $storeIds = [], bool $unused = false): void
     {
